requirements.txt
discord.py[voice]>=2.3.2
python-dotenv
sqlalchemy
psycopg2-binary
pandas
openpyxl
apscheduler